sdk/python/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
absl-py==0.11.0
    # via -r requirements.in
attrs==21.2.0
    # via jsonschema
cachetools==4.2.4
    # via google-auth
certifi==2021.5.30
    # via
    #   kfp-server-api
    #   kubernetes
    #   requests
charset-normalizer==2.0.6
    # via requests
click==8.0.3
    # via
    #   -r requirements.in
    #   typer
cloudpickle==2.0.0
    # via -r requirements.in
deprecated==1.2.13
    # via -r requirements.in
docstring-parser==0.11
    # via -r requirements.in
fire==0.4.0
    # via -r requirements.in
google-api-core==2.6.0
    # via
    #   -r requirements.in
    #   google-cloud-core
    #   google-cloud-storage
google-auth==1.35.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
    #   kubernetes
google-cloud-core==2.1.0
    # via google-cloud-storage
google-cloud-storage==2.2.1
    # via -r requirements.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
idna==3.2
    # via requests
jsonschema==3.2.0
    # via -r requirements.in
kfp-pipeline-spec==0.1.14
    # via -r requirements.in
kfp-server-api==2.0.0a0
    # via -r requirements.in
kubernetes==18.20.0
    # via -r requirements.in
oauthlib==3.1.1
    # via requests-oauthlib
protobuf==3.17.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   kfp-pipeline-spec
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pydantic==1.8.2
    # via -r requirements.in
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   kfp-server-api
    #   kubernetes
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   kubernetes
requests-oauthlib==1.3.0
    # via kubernetes
requests-toolbelt==0.9.1
    # via -r requirements.in
requests==2.26.0
    # via
    #   google-api-core
    #   google-cloud-storage
    #   kubernetes
    #   requests-oauthlib
    #   requests-toolbelt
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   absl-py
    #   fire
    #   google-auth
    #   google-cloud-storage
    #   jsonschema
    #   kfp-server-api
    #   kubernetes
    #   protobuf
    #   python-dateutil
strip-hints==0.1.10
    # via -r requirements.in
tabulate==0.8.9
    # via -r requirements.in
termcolor==1.1.0
    # via fire
typer==0.4.0
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via pydantic
uritemplate==3.0.1
    # via -r requirements.in
urllib3==1.26.7
    # via
    #   kfp-server-api
    #   kubernetes
    #   requests
websocket-client==1.2.1
    # via kubernetes
wheel==0.37.0
    # via strip-hints
wrapt==1.13.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
